requirements.txt
langchain==1.0.2
langchain-core==1.0.1
langchain-deepseek==1.0.0a1
langchain-huggingface==1.0.0
langchain-mcp-adapters==0.1.11
mcp==1.19.0
pydantic==2.11.9
python-dotenv==1.2.1
typing_extensions==4.15.0
